trying out usize

+++ src/composer.rs
@@ -2,7 +2,11 @@ use crate::*;
 
 /// # Safety
 /// pippenger must point to a valid Pippenger object
-pub unsafe fn get_solidity_verifier(g2_ptr: &[u8], vk_ptr: &[u8], output_buf: *mut *mut u8) -> u32 {
+pub unsafe fn get_solidity_verifier(
+    g2_ptr: &[u8],
+    vk_ptr: &[u8],
+    output_buf: *mut *mut u8,
+) -> usize {
     acir_proofs_get_solidity_verifier(
         g2_ptr.as_ptr() as *const u8,
         vk_ptr.as_ptr() as *const u8,
